+++ components/Write/Layout.tsx
@@ -1,55 +1,18 @@
-import React, { useCallback, useEffect } from 'react';
-import { useRouter } from 'next/router';
-import { useRecoilValue, useResetRecoilState, useSetRecoilState } from 'recoil';
+import React, { useEffect } from 'react';
+import { useRecoilValue, useResetRecoilState } from 'recoil';
 import { progressStepStateAtom } from '@/store/progressStep/atom';
 import { createPostRequestState } from '@/store/post/atom';
-import { CommonAppBar, CommonIconButton, CommonProgress } from '../Common';
-import styled from 'styled-components';
-import theme from '@/styles/theme';
+import { CommonProgress } from '../Common';
+import AppBar from './AppBar';
 
 interface LayoutProps {
   children: React.ReactElement;
 }
 
 const Layout = ({ children }: LayoutProps) => {
-  const router = useRouter();
   const progressStep = useRecoilValue(progressStepStateAtom);
-  const setPrevProgressStep = useSetRecoilState(progressStepStateAtom);
   const resetcreatePostRequestState = useResetRecoilState(createPostRequestState);
   const resetProgressStepState = useResetRecoilState(progressStepStateAtom);
-
-  const onClickGoBack = useCallback(() => {
-    if (progressStep === 1) {
-      router.back();
-      return;
-    }
-    setPrevProgressStep((prev) => prev - 1);
-  }, [progressStep, router, setPrevProgressStep]);
-
-  const renderAppBar = () => {
-    if (progressStep === 1) {
-      return (
-        <CommonAppBar>
-          <CommonAppBar.Left>
-            <CommonIconButton iconName="left" alt="이전" onClick={onClickGoBack} />
-          </CommonAppBar.Left>
-        </CommonAppBar>
-      );
-    }
-    if (progressStep === 4) {
-      return <EmptyAppBar />;
-    }
-    return (
-      <CommonAppBar>
-        <CommonAppBar.Left>
-          <CommonIconButton iconName="left" alt="이전" onClick={onClickGoBack} />
-        </CommonAppBar.Left>
-        <CommonAppBar.Right>
-          <CancelText onClick={() => router.back()}>취소</CancelText>
-        </CommonAppBar.Right>
-      </CommonAppBar>
-    );
-  };
 
   useEffect(() => {
     return () => {
@@ -60,7 +23,7 @@ const Layout = ({ children }: LayoutProps) => {
 
   return (
     <>
-      {renderAppBar()}
+      <AppBar />
       <CommonProgress step={progressStep} />
       {children}
     </>
@@ -68,15 +31,3 @@ const Layout = ({ children }: LayoutProps) => {
 };
 
 export default Layout;
-
-const CancelText = styled.div`
-  ${theme.fonts.h6}
-  line-height: 1.68rem;
-  color: ${theme.colors.white};
-  cursor: pointer;
-`;
-
-const EmptyAppBar = styled.div`
-  width: 100%;
-  height: 4.6rem;
-`;
